Fix Cron (#5)
The Cron was very broken before 🤦 This is something I kinda knew, but
though the issue was more in Tokio/infra land and never dug in much.

But it turns out the issue was that my worker code was very badly
written! It marked the `last_enqueued_at` on EACH tick, even if a job
wasn't enqueued!

Now that is fixed, and we have more tests!